fix: remove the Executor pool
The Executor "pool" doesn't have the option to create unlimited number of threads, it's always a single thread or "n" threads. This commit is a potential candidate to be reverted, because the pool controls the threads and resources very well. I'll have to do some stress tests on this (connecting with a lot of clients) to check if the use of resources and performance will not be compromised.